feat: implement v4.00 - unlock SuperAdmin Settings, Admin Accomplishments & Ratings, Faculty IPCR Form
Cherry-picked onto the v3.00 snapshot. Includes b349d08 so the newly
unlocked IPCR Form submits correctly and shows the immediate supervisor.

+++ database/csu_piat_aopcr_ipcr.sql
@@ -263,7 +263,7 @@ CREATE TABLE `ipcr_items` (
   `function_type` enum('core','strategic','support') NOT NULL,
   `success_indicator` text DEFAULT NULL,
   `accomplishment` text DEFAULT NULL,
-  `rating` tinyint(4) DEFAULT NULL CHECK (`rating` between 1 and 5),
+  `rating` decimal(2,1) DEFAULT NULL CHECK (`rating` between 1 and 5),
   `remarks` varchar(200) DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
